Replaced always-active Sandbox button w/ disabled button when an objective is selected
Original implementation involved toggling the selected objective to ensure that the sandbox button always went to sandbox mode, but it was discovered that the act of toggling the objective (on or off) would include resetting the game rules to the recommended settings based on the objective's definition, which is undesirable behavior. The objective selection toggling has been deleted, but this requires the sandbox button to not be selectable when an objective is selected by the player, else they could select an objective, press Sandbox, then be playing the selected objective.